+++ xenon-common/src/main/java/com/vmware/xenon/common/Operation.java
@@ -302,7 +302,17 @@ public class Operation implements Cloneable {
         /**
          * Set by client to disable default logging of operation failures
          */
-        FAILURE_LOGGING_DISABLED
+        FAILURE_LOGGING_DISABLED,
+        /**
+         * Set by a local {@code ServiceRequestListener} instance indicating the operation
+         * originated outside the process / host
+         */
+        REMOTE,
+        /**
+         * The operation exceeded the rate limit associated with it logical context
+         * (authorization subject by default)
+         */
+        RATE_LIMITED
     }
 
     public static class SerializedOperation extends ServiceDocument {
@@ -765,13 +775,12 @@ public class Operation implements Cloneable {
     }
 
     public boolean isRemote() {
-        return this.remoteCtx != null && this.remoteCtx.socketCtx != null;
+        return this.options.contains(OperationOption.REMOTE) ||
+                (this.remoteCtx != null && this.remoteCtx.socketCtx != null);
     }
 
     public Operation forceRemote() {
-        allocateRemoteContext();
-        this.remoteCtx.socketCtx = new SocketContext();
-        return this;
+        return toggleOption(OperationOption.REMOTE, true);
     }
 
     public AuthorizationContext getAuthorizationContext() {
@@ -782,7 +791,7 @@ public class Operation implements Cloneable {
      * Sets (overwrites) the authorization context of this operation.
      *
      * The visibility of this method is intentionally package-local. It is intended to
-     * only be called by functions in this package, so that we can apply whitelisting
+     * only be called by functions in this package, so that we can apply white listing
      * to limit the set of services that is able to set it.
      *
      * @param ctx the authorization context to set.